car/longtail/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
scikit-learn>=1.3.0
tqdm>=4.65.0
requests>=2.31.0
